backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-environ==0.12.0
drf-spectacular==0.27.2
django-cors-headers==4.7.0
djangorestframework-simplejwt==5.5.0
# Database
psycopg2==2.9.10
psycopg2-binary==2.9.10
dj-database-url==2.3.0
# Image
Pillow==11.1.0
# Testing
ipython==8.32.0
